Code Source
These files simulate functionalities applied to the project. They can be used to simulate an SFC in three different topologies, sequential, hybrid and parallel. In each test the location and internal variables need to be configured.

A brief description of the files:
RTBilling: Service with just one function used to count the number of bytes without a sequence of messages between the source and destination.

RTCripter: Service with just one function used to encrypt messages between the source and destination.

RTLogging: Service with just one function used to check the password and save the message of a sequence of messages between the source and destination.

RTTAnalysis: Service with just one function used to account for the different types of characters present in a message in a sequence of messages between the source and destination.

SendDatagram: Client program that sends a message.

ServerUDP: Simple UDP server that receives messages from the client.